requirements.txt
PyMuPDF
bert-extractive-summarizer
transformers
torch
sentence-transformers
pytesseract
Pillow
scikit-learn
streamlit
